// Fontend/TNT.N8.Web/TNT.N8.Web/ClientApp/src/app/shared/models/cauHinhHeSoKhuyenKhich.model.ts
export class CauHinhHeSoKhuyenKhichModel {
    id: string | null;
    chucVuId: number | null;
    chucVu: string;
    dichVu: string;
    loaiThuongId: number | null;
    loaiThuong: string;
    giaTriThuong: number | null;
    tuNgay: string | null;
    denNgay: string | null;
    dieuKienId: number | null;
    dieuKien: string;
    type: number | null;
    doiTuongApDungId: number | null;
    doiTuongApDung: string | null;
    giaTriTu: number | null;
    giaTriDen: number | null;
    parentId: string | null;
    createdDate: string | null;
    createdById: string | null;
    updatedDate: string | null;
    updatedById: string | null;
    tenantId: string | null;
}