// Fontend/TNT.N8.Web/TNT.N8.Web/ClientApp/src/app/product/models/product.model.ts
export class OptionsEntityModel {
  id: string;
  parentId: string;
  categoryId: string;
  name: string;
  nameCustom: string;
  description: string;
  note: string;
  enableNumber: boolean | null;
  isShowQuantity: boolean;
  isActive: boolean;
  createdById: string;
  createdDate: string;
  updatedById: string;
  updatedDate: string;
  tenantId: string;
  price: number;
  ServiceTypeName: string;
  servicePacketId: string;
  isExtend: boolean;
}
// ...
export class EmployeeEntityModel {
  employeeId: string;
  mission: string;
  employeeName: string;
  employeeCode: string;
  OrganizationId: string;
  PositionId: string;
  EmployeeCodeName: string;
  roleId : string | null;
  isDisable : boolean | null;
  value : string | null;
  chucVuId: number;
}
// ...
export class TrangThaiGeneral {
  value: number;
  valueText: string;
  name: string;
  objectRole : TrangThaiGeneral;
  listEmployeeEntityModel : EmployeeEntityModel[];
  isEdit : boolean | null;
  categoryId : string | null;
}